Move `DefaultInputDrum` to correct location
